+++ src/modules/client/pages/ni/courses/LearnersDirectory.vue
@@ -24,18 +24,12 @@
 </template>
 
 <script>
-import get from 'lodash/get';
 import { mapGetters } from 'vuex';
 import Users from '@api/Users';
 import TableList from '@components/table/TableList';
 import DirectoryHeader from '@components/DirectoryHeader';
 import LearnerCreationModal from '@components/courses/LearnerCreationModal';
-import {
-  formatPhoneForPayload,
-  removeEmptyProps,
-  clear,
-} from '@helpers/utils';
-import { NotifyPositive, NotifyNegative, NotifyWarning } from '@components/popup/notify';
+import { NotifyPositive, NotifyNegative } from '@components/popup/notify';
 import { userMixin } from '@mixins/userMixin';
 import { learnerDirectoryMixin } from '@mixins/learnerDirectoryMixin';
 
@@ -51,40 +45,7 @@ export default {
   computed: {
     ...mapGetters({ company: 'main/getCompany' }),
   },
-  async created () {
-    await this.getLearnerList(this.company._id);
-  },
   methods: {
-    resetAddLearnerForm () {
-      this.firstStep = true;
-      this.newLearner = { ...clear(this.newLearner) };
-      this.$v.newLearner.$reset();
-    },
-    async nextStepLearnerCreationModal () {
-      try {
-        this.$v.newLearner.$touch();
-        if (this.$v.newLearner.local.email.$error) return NotifyWarning('Champ invalide.');
-
-        this.learnerCreationModalLoading = true;
-        const userInfo = await Users.exists({ email: this.newLearner.local.email });
-
-        if (!userInfo.exists) return this.goToCreationStep();
-        if (!userInfo.user.company && userInfo.user._id) return this.updateLearner(userInfo.user._id);
-
-        return get(userInfo, 'user.company') === this.company._id
-          ? NotifyWarning('L\'apprenant(e) est déjà ajouté(e).')
-          : NotifyNegative('L\'apprenant(e) n\'est pas relié(e) à cette structure.');
-      } catch (e) {
-        console.error(e);
-        NotifyNegative('Erreur lors de l\'ajout de l\'apprenant(e).');
-      } finally {
-        this.learnerCreationModalLoading = false;
-      }
-    },
-    goToCreationStep () {
-      this.firstStep = false;
-      this.$v.newLearner.$reset();
-    },
     async updateLearner (userId) {
       try {
         await Users.updateById(userId, { company: this.company._id });
@@ -96,34 +57,6 @@ export default {
         console.error(e);
         if (e.status === 409) return NotifyNegative(e.data.message);
         NotifyNegative('Erreur lors de l\'ajout de l\'apprenant(e).');
-      }
-    },
-    formatUserPayload () {
-      const payload = removeEmptyProps(this.newLearner);
-      if (get(payload, 'contact.phone')) payload.contact.phone = formatPhoneForPayload(this.newLearner.contact.phone);
-
-      return ({ ...payload, company: this.company._id });
-    },
-    async createLearner () {
-      try {
-        this.learnerCreationModalLoading = true;
-        this.$v.newLearner.$touch();
-        if (this.$v.newLearner.$error) return NotifyWarning('Champ(s) invalide(s).');
-
-        const payload = await this.formatUserPayload();
-        await Users.create(payload);
-        NotifyPositive('Apprenant(e) ajouté(e) avec succès.');
-
-        await this.sendWelcome();
-
-        this.learnerCreationModal = false;
-        await this.getLearnerList(this.company._id);
-      } catch (e) {
-        console.error(e);
-        if (e.status === 409) return NotifyNegative(e.data.message);
-        NotifyNegative('Erreur lors de l\'ajout de l\' apprenant(e).');
-      } finally {
-        this.learnerCreationModalLoading = false;
       }
     },
   },
